# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backend server for alfarpos.",
        "title": "AlfarPOS BackEnd",
        "contact": {
            "name": "Almantera Tiantana Al Faruqi",
            "url": "https://twitter.com/almanalfaruq",
            "email": "alman.alfaruq@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Get Products based on query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get Products based on query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unit or category",
                        "name": "searchBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If this empty, it will fetch all products",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return array of product",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/product.Product"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/products/code/{code}": {
            "get": {
                "description": "Get Product based on code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get Product based on code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code of the product",
                        "name": "coded",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a product",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/product.Product"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/products/id/{id}": {
            "get": {
                "description": "Get Product based on id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get Product based on id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the product",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a product",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/product.Product"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/products/ids/{ids}": {
            "get": {
                "description": "Get Multiple Product based on id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get Multiple Product based on ids",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ids of the product",
                        "name": "ids",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a product",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/product.Product"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/products/upload_excel": {
            "post": {
                "description": "Upload products from excel file (*.xlsx). First row should be the header with these column in order:\nCode (Barcode), Product Name, Sell Price, Quantity, Category Name, Buy Price, and Unit Name (pcs, bottle, etc.)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Upload products from excel file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the sheet",
                        "name": "sheetName",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return array of product",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/product.Product"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Category"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "product.Product": {
            "type": "object",
            "properties": {
                "buy_price": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "code": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "discount": {
                    "$ref": "#/definitions/sql.NullFloat64"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "type": "string",
                    "example": "http://localhost/image/image.jpg"
                },
                "is_open_price": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "example": "product name"
                },
                "product_prices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.ProductPrice"
                    }
                },
                "quantity": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "related_products": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sell_price": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "unit": {
                    "$ref": "#/definitions/model.Unit"
                },
                "unit_id": {
                    "type": "integer",
                    "example": 2
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success getting all products"
                }
            }
        },
...
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}